auth-11
- функция  записи шопинг листа из БД в локалсторедж при авторизации,
- проверка имени на пробелы при регистрации,
- исправление бордера кнопки в цсс
